fix parsing numbers in EntityUtils.setValue()
(cherry picked from commit 4353fb0a8b79135a3f7d551c9b7c93850fc1db85)

+++ src/platform-kit/instance-pipeline/EntityUtils.ts
@@ -45,7 +45,7 @@ export class EntityUtils {
 				entityRecord[key] = parsedValue.asByteArray()
 				break
 			case ValueTypeEnum.Number:
-				entityRecord[key] = assertNotNaN(parseInt(parsedValue.asString()), `Non-numeric string for attribute: ${modelValue.name}`).toString()
+				entityRecord[key] = assertNotNaN(parseFloat(parsedValue.asString()), `Non-numeric string for attribute: ${modelValue.name}`).toString()
 				break
 			case ValueTypeEnum.String:
 			case ValueTypeEnum.CompressedString:
